geminicli/requirements.txt
click
httpx
pyyaml
